requirements.txt
discord-components
discord-py-interactions
discord-typings
discord.py